refactor(utils): move some functions from utils to items.py
Also, updated the imports and temporarily commented out the corresponding functions in utils
